move snapshot serialization tests to kernel tests
