+++ tensorflow/python/data/experimental/kernel_tests/serialization/BUILD
@@ -123,25 +123,6 @@ tf_py_test(
 )
 
 tf_py_test(
-    name = "snapshot_dataset_serialization_test",
-    size = "medium",
-    srcs = ["snapshot_dataset_serialization_test.py"],
-    shard_count = 3,
-    tags = [
-        "no_oss",
-        "no_pip",
-        "no_windows",
-    ],
-    deps = [
-        "//tensorflow/python:client_testlib",
-        "//tensorflow/python/data/kernel_tests:checkpoint_test_base",
-        "//tensorflow/python/data/kernel_tests:test_base",
-        "//tensorflow/python/data/ops:dataset_ops",
-        "@absl_py//absl/testing:parameterized",
-    ],
-)
-
-tf_py_test(
     name = "dataset_constructor_serialization_test",
     size = "medium",
     srcs = ["dataset_constructor_serialization_test.py"],
